web/requirements.txt
django
mysqlclient
python-dotenv
Pillow
django-otp
django-environ
django-two-factor-auth[phonenumbers]
qrcode
pyotp
django-otp
djangorestframework
django-filter
openpyxl
python-telegram-bot
whisper
torch
redis
django-redis
requests